requirements.txt
Flask 
Flask-SQLAlchemy 
psycopg2-binary
python-dotenv
Flask-Migrate
pytest 
pytest-flask
pytest-mock